requirements.txt
googlemaps
google-generativeai
mcp
python-dotenv
requests
folium
flask
